# openapi.json
{"openapi":"3.1.0","info":{"title":"Cron Schedule API","description":"Vypočítá nejbližší budoucí (nebo minulé) termíny podle CRON výrazu.","version":"1.0.0"},"paths":{"/next":{"get":{"summary":"Next Runs","description":"Vrátí pole ISO 8601 časů příštích spuštění.","operationId":"next_runs_next_get","parameters":[{"name":"expr","in":"query","required":true,"schema":{"type":"string","description":"CRON výraz – např. */15 * * * *","title":"Expr"},"description":"CRON výraz – např. */15 * * * *"},{"name":"count","in":"query","required":false,"schema":{"type":"integer","maximum":100,"exclusiveMinimum":0,"description":"Počet termínů (1-100)","default":5,"title":"Count"},"description":"Počet termínů (1-100)"},{"name":"tzname","in":"query","required":false,"schema":{"type":"string","description":"Časové pásmo, default UTC","default":"UTC","title":"Tzname"},"description":"Časové pásmo, default UTC"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/previous":{"get":{"summary":"Previous Runs","description":"Vrátí pole ISO 8601 časů minulých spuštění.","operationId":"previous_runs_previous_get","parameters":[{"name":"expr","in":"query","required":true,"schema":{"type":"string","description":"CRON výraz","title":"Expr"},"description":"CRON výraz"},{"name":"count","in":"query","required":false,"schema":{"type":"integer","maximum":100,"exclusiveMinimum":0,"default":5,"title":"Count"}},{"name":"tzname","in":"query","required":false,"schema":{"type":"string","default":"UTC","title":"Tzname"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/validate":{"get":{"summary":"Validate","description":"Jen zkontroluje syntaxi CRON výrazu.","operationId":"validate_validate_get","parameters":[{"name":"expr","in":"query","required":true,"schema":{"type":"string","description":"CRON výraz k validaci","title":"Expr"},"description":"CRON výraz k validaci"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Health-check","description":"Jednoduchá sonda pro CI / Render.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}